elif added

+++ Course-5-flow-control.py
@@ -50,7 +50,7 @@ else:
 # Make sure to run your program multiple times and test it with values for each 5 of the different grades
 # to make sure that the correct output is displayed for any value entered as a student's score.
 
-verScore = float(input("Please insert the student's scores."))
+verScore = int(input("Please insert the student's scores."))
 
 if verScore >= 90:
     print("Your grade is A")
@@ -65,3 +65,7 @@ else:
                 print("Your grade is D")
             else:
                 print("Your grade is F")
+
+
+# elif statement
+
